LayoutTests/platform/ipad/fast/viewport/ios/device-width-viewport-after-changing-view-scale-expected.txt

setViewScale(0.50)
window size: [1620, 2120]
square size: [162, 212]
zoom scale: 0.50

setViewScale(0.75)
window size: [1080, 1413]
square size: [108, 141]
zoom scale: 0.75

setViewScale(1.00)
window size: [810, 1060]
square size: [81, 106]
zoom scale: 1.00

setViewScale(1.25)
window size: [648, 848]
square size: [65, 85]
zoom scale: 1.25

setViewScale(1.50)
window size: [540, 707]
square size: [54, 71]
zoom scale: 1.50

setViewScale(1.25)
window size: [648, 848]
square size: [65, 85]
zoom scale: 1.25

setViewScale(1.00)
window size: [810, 1060]
square size: [81, 106]
zoom scale: 1.00

setViewScale(0.75)
window size: [1080, 1413]
square size: [108, 141]
zoom scale: 0.75

setViewScale(0.50)
window size: [1620, 2120]
square size: [162, 212]
zoom scale: 0.50
